Updates for 0.0.3 including quickstart fixes.
