+++ MAINTAINERS.md
@@ -1,4 +1,5 @@
 # MAINTAINERS
 
-Mark Taylor  (ibmmqmet) - marke_taylor@uk.ibm.com
-Kate Stanley (katheris) - katheris@uk.ibm.com
+- Mark Taylor  (ibmmqmet) - marke_taylor@uk.ibm.com
+- Andrew Schofield (AndrewJSchofield) - andrew_schofield@uk.ibm.com
+- Kate Stanley (katheris) - katheris@uk.ibm.com
