+++ config/elastic-sink.properties
@@ -12,24 +12,25 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-name=es-sink
-connector.class=com.ibm.eventstreams.connect.essink.ESSinkConnector
+name=elastic-sink
+connector.class=com.ibm.eventstreams.connect.elasticsink.ElasticSinkConnector
 
 # You can increase this for higher throughput to match the number of topic partitions
 tasks.max=1
 
 # The list of source Kafka topics
-topics=<TOPIC>
+topics=
 
 # The hostname:port to connect to Elasticsearch.
 es.connection=localhost:9200
 
 # Conversion of the Kafka message to a document is done via this class
-es.document.builder=com.ibm.eventstreams.connect.essink.builders.BaseDocumentBuilder
+es.document.builder=com.ibm.eventstreams.connect.elasticsink.builders.JsonDocumentBuilder
 # Creation of the index for the document is done via this class
-es.index.builder=com.ibm.eventstreams.connect.essink.builders.BaseIndexBuilder
+es.index.builder=com.ibm.eventstreams.connect.elasticsink.builders.DefaultIndexBuilder
 # Creation of the identifier for the document is done via this class
-es.identifier.builder=com.ibm.eventstreams.connect.essink.builders.DefaultIdentifierBuilder
+es.identifier.builder=com.ibm.eventstreams.connect.elasticsink.builders.DefaultIdentifierBuilder
+#es.identifier.builder=com.ibm.eventstreams.connect.elasticsink.builders.KeyIdentifierBuilder
 
 # Authentication to Elasticsearch. Defaults are empty
 #es.user.name=
@@ -56,11 +57,6 @@ es.identifier.builder=com.ibm.eventstreams.connect.essink.builders.DefaultIdenti
 #es.max.failures=5
  
 # These converters control conversion of data between the internal Kafka Connect representation and the messages in Kafka.
-# key.converter=org.apache.kafka.connect.converters.ByteArrayConverter
 key.converter=org.apache.kafka.connect.storage.StringConverter
-# key.converter=org.apache.kafka.connect.json.JsonConverter
-
-# value.converter=org.apache.kafka.connect.converters.ByteArrayConverter
-# value.converter=org.apache.kafka.connect.storage.StringConverter
 value.converter=org.apache.kafka.connect.json.JsonConverter
 value.converter.schemas.enable=false
