--- config/es-sink.json
@@ -1,16 +1,0 @@
-{
-    "name": "es-sink",
-    "config":
-    {
-        "connector.class": "com.ibm.eventstreams.connect.essink.ESSinkConnector",
-        "tasks.max": "1",
-        "topics": "<TOPIC>",
-        "es.connection": "localhost:9200",
-        "es.document.builder": "com.ibm.eventstreams.connect.essink.builders.BaseDocumentBuilder",
-        "es.index.builder": "com.ibm.eventstreams.connect.essink.builders.BaseIndexBuilder",
-        "es.identifier.builder": "com.ibm.eventstreams.connect.essink.builders.DefaultIdentifierBuilder",
-        "key.converter": "org.apache.kafka.connect.storage.StringConverter",
-        "value.converter": "org.apache.kafka.connect.json.JsonConverter",
-        "value.converter.schemas.enable": "false"
-    }
-}
