+++ pom.xml
@@ -18,10 +18,10 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.ibm.eventstreams.connect</groupId>
-  <artifactId>kafka-connect-es-sink</artifactId>
+  <artifactId>kafka-connect-elastic-sink</artifactId>
   <packaging>jar</packaging>
-  <version>0.0.2-SNAPSHOT</version>
-  <name>kafka-connect-es-sink</name>
+  <version>0.0.3-SNAPSHOT</version>
+  <name>kafka-connect-elastic-sink</name>
   <organization>
     <name>IBM Corporation</name>
     <url>http://ibm.com</url>
@@ -45,13 +45,13 @@
     <dependency>
       <groupId>org.apache.kafka</groupId>
       <artifactId>connect-api</artifactId>
-      <version>2.4.0</version>
+      <version>2.5.0</version>
       <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>org.apache.kafka</groupId>
       <artifactId>connect-json</artifactId>
-      <version>2.4.0</version>
+      <version>2.5.0</version>
       <scope>provided</scope>
     </dependency>
