+++ quickstart/Dockerfile
@@ -1,5 +1,5 @@
-#FROM ibmcom/eventstreams-kafka-ce-icp-linux-amd64:2019.2.1-3a2f93e as kafka
-FROM ibmcom/eventstreams-kafkaconnect:latest as kafka
+FROM strimzi/kafka:latest-kafka-2.5.0 as kafka
+
 FROM ibmjava:8-jre
 
 RUN addgroup --gid 5000 --system esgroup && \
@@ -13,12 +13,12 @@ RUN mkdir /opt/kafka/logs && chown esuser:esgroup /opt/kafka/logs
 
 # These are the key files needed from this repository to get something running. The hostnames in the
 # properties files reflect the docker-compose names
-COPY --chown=esuser:esgroup target/kafka-connect-es-sink-*-jar-with-dependencies.jar /opt/kafka/libs/
-COPY --chown=esuser:esgroup quickstart/es-sink-qs.properties /opt/kafka/config
+COPY --chown=esuser:esgroup target/kafka-connect-elastic-sink-*-jar-with-dependencies.jar /opt/kafka/libs/
+COPY --chown=esuser:esgroup quickstart/elastic-sink-qs.properties /opt/kafka/config
 COPY --chown=esuser:esgroup quickstart/connect-standalone-qs.properties /opt/kafka/config/
 
 WORKDIR /opt/kafka
 
 USER esuser
 
-ENTRYPOINT ["./bin/connect-standalone.sh", "config/connect-standalone-qs.properties", "config/es-sink-qs.properties"]
+ENTRYPOINT ["./bin/connect-standalone.sh", "config/connect-standalone-qs.properties", "config/elastic-sink-qs.properties"]
