+++ quickstart/elastic-sink-qs.properties
@@ -12,21 +12,21 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-name=es-sink
-connector.class=com.ibm.eventstreams.connect.essink.ESSinkConnector
+name=elastic-sink
+connector.class=com.ibm.eventstreams.connect.elasticsink.ElasticSinkConnector
 
 # The list of source Kafka topics
-topics=ESTESTSINK
+topics=ESSINKTEST
 
 # The hostname:port to connect to Elasticsearch.
 es.connection=elastic:9200
 
 # Conversion of the Kafka message to a document is done via this class
-es.document.builder=com.ibm.eventstreams.connect.essink.builders.BaseDocumentBuilder
+es.document.builder=com.ibm.eventstreams.connect.elasticsink.builders.JsonDocumentBuilder
 # Creation of the index for the document is done via this class
-es.index.builder=com.ibm.eventstreams.connect.essink.builders.BaseIndexBuilder
+es.index.builder=com.ibm.eventstreams.connect.elasticsink.builders.DefaultIndexBuilder
 # Creation of the identifier for the document is done via this class
-es.identifier.builder=com.ibm.eventstreams.connect.essink.builders.DefaultIdentifierBuilder
+es.identifier.builder=com.ibm.eventstreams.connect.elasticsink.builders.DefaultIdentifierBuilder
 
 # These converters control conversion of data between the internal Kafka Connect representation and the messages in Kafka.
 key.converter=org.apache.kafka.connect.storage.StringConverter
