+++ src/main/java/com/ibm/eventstreams/connect/elasticsink/builders/IdentifierBuilder.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.ibm.eventstreams.connect.essink.builders;
+package com.ibm.eventstreams.connect.elasticsink.builders;
 
 import java.util.Map;
 
@@ -23,28 +23,28 @@ import org.apache.kafka.connect.sink.SinkRecord;
  * Builds document identifiers from Kafka Connect SinkRecords.
  */
 public interface IdentifierBuilder {
-    /**
-     * Configure this class.
-     *
-     * @param props initial configuration
-     *
-     * @throws ConnectException   Operation failed and connector should stop.
-     */
-    default void configure(Map<String, String> props) {}
+  /**
+   * Configure this class.
+   *
+   * @param props initial configuration
+   *
+   * @throws ConnectException   Operation failed and connector should stop.
+   */
+  default void configure(Map<String, String> props) {}
 
-    /**
-     * Convert a Kafka Connect SinkRecord into a document identifier.
-     *
-     * @param record             the Kafka Connect SinkRecord
-     *
-     * @return the document identifier
-     */
-    String fromSinkRecord(SinkRecord record);
+  /**
+   * Convert a Kafka Connect SinkRecord into a document identifier.
+   *
+   * @param record             the Kafka Connect SinkRecord
+   *
+   * @return the document identifier
+   */
+  String fromSinkRecord(SinkRecord record);
 
-    /**
-     * Whether document identifiers generated by this builder are unique.
-     * 
-     * @return true if unique, false if not
-     */
-    boolean isUnique();
+  /**
+   * Whether document identifiers generated by this builder are unique.
+   * 
+   * @return true if unique, false if not
+   */
+  boolean isUnique();
 }
