+++ src/main/java/com/ibm/eventstreams/connect/elasticsink/builders/KeyIdentifierBuilder.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.ibm.eventstreams.connect.essink.builders;
+package com.ibm.eventstreams.connect.elasticsink.builders;
 
 import org.apache.kafka.connect.errors.ConnectException;
 import org.apache.kafka.connect.sink.SinkRecord;
@@ -22,29 +22,29 @@ import org.apache.kafka.connect.sink.SinkRecord;
  * Builds document identifiers from Kafka Connect SinkRecords.
  */
 public class KeyIdentifierBuilder implements IdentifierBuilder {
-    /**
-     * Convert a Kafka Connect SinkRecord into a document identifier.
-     *
-     * @param record             the Kafka Connect SinkRecord
-     *
-     * @return the document identifier
-     */
-    @Override
-    public String fromSinkRecord(SinkRecord record) {
-      if (record.key() == null) {
-          throw new ConnectException("Key must be present to use as document ID.");
-      }
-      return record.key().toString();        
+  /**
+   * Convert a Kafka Connect SinkRecord into a document identifier.
+   *
+   * @param record             the Kafka Connect SinkRecord
+   *
+   * @return the document identifier
+   */
+  @Override
+  public String fromSinkRecord(SinkRecord record) {
+    if (record.key() == null) {
+      throw new ConnectException("Key must be present to use as document ID.");
     }
+    return record.key().toString();        
+  }
 
-    /**
-     * Whether document identifiers generated by this builder are unique.
-     * 
-     * @return true if unique, false if not
-     */
-    @Override
-    public boolean isUnique() {
-        // Kafka record keys are not guaranteed to be unique
-        return false;
-    }
+  /**
+   * Whether document identifiers generated by this builder are unique.
+   * 
+   * @return true if unique, false if not
+   */
+  @Override
+  public boolean isUnique() {
+    // Kafka record keys are not guaranteed to be unique
+    return false;
+  }
 }
